Fix MSVS build

+++ src/mandelbulber/fractal.cpp
@@ -20,7 +20,6 @@
 
 #include "fractal.h"
 #include "algebra.cpp"
-
 /**
  * Compute the fractal at the point, in one of the various modes
  *
@@ -1178,7 +1177,7 @@ double Compute(CVector3 z, const sFractal &par, int *iter_count)
 					distance = z.Length();
 					if (distance > 1e15)
 					{
-						distance = (L - log(log(r) / log(N)) / log(p))/100.0;
+						distance = (L - log(log(r) / log((double)N)) / log(p))/100.0;
 						break;
 					}
 				}
